+++ src/test/php/lang/ast/unittest/parse/VariablesTest.class.php
@@ -36,14 +36,4 @@ class VariablesTest extends ParseTest {
       '$a[0];'
     );
   }
-
-  /** @deprecated */
-  #[Test]
-  public function string_offset() {
-    $this->assertParsed(
-      [new OffsetExpression(new Variable('a', self::LINE), new Literal('0', self::LINE), self::LINE)],
-      '$a{0};'
-    );
-    \xp::gc();
-  }
 }